+++ R/OptimizerCmaes.R
@@ -78,6 +78,7 @@ OptimizerCmaes = R6Class("OptimizerCmaes",
         param_classes = "ParamDbl",
         properties = "single-crit",
         packages = "adagio",
+        label = "Covariance Matrix Adaptation Evolution Strategy",
         man = "bbotk::mlr_optimizers_cmaes"
       )
     }