+++ R/TerminatorRunTime.R
@@ -1,11 +1,10 @@
-#' @title Terminator that stops according to the run time
+#' @title Run Time Terminator
 #'
 #' @name mlr_terminators_run_time
 #' @include Terminator.R
 #'
 #' @description
-#' Class to terminate the optimization after the optimization process took a
-#' number of seconds on the clock.
+#' Class to terminate the optimization after the optimization process took a number of seconds on the clock.
 #'
 #' @templateVar id run_time
 #' @template section_dictionary_terminator
@@ -13,7 +12,7 @@
 #' @section Parameters:
 #' \describe{
 #' \item{`secs`}{`numeric(1)`\cr
-#' Maximum allowed time, in seconds, default is 100.}
+#'   Maximum allowed time, in seconds, default is 100.}
 #' }
 #'
 #' @note
@@ -21,13 +20,16 @@
 #' done by the [Optimizer].
 #'
 #' @family Terminator
+#'
 #' @template param_archive
+#'
 #' @export
 #' @examples
 #' trm("run_time", secs = 1800)
 TerminatorRunTime = R6Class("TerminatorRunTime",
   inherit = Terminator,
   public = list(
+
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
@@ -35,13 +37,18 @@ TerminatorRunTime = R6Class("TerminatorRunTime",
         secs = p_dbl(lower = 0, default = 30)
       )
       param_set$values$secs = 30
-      super$initialize(param_set = param_set, properties = c("single-crit", "multi-crit"))
-      self$unit = "seconds"
+      super$initialize(
+        param_set = param_set,
+        properties = c("single-crit", "multi-crit"),
+        unit = "seconds",
+        label = "Run Time",
+        man = "bbotk::mlr_terminators_run_time")
     },
 
     #' @description
     #' Is `TRUE` iff the termination criterion is positive, and `FALSE`
     #' otherwise.
+    #'
     #' @return `logical(1)`.
     is_terminated = function(archive) {
       assert_r6(archive, "Archive")