+++ R/mlr_optimizers.R
@@ -4,28 +4,37 @@
 #' @format [R6::R6Class] object inheriting from [mlr3misc::Dictionary].
 #'
 #' @description
-#' A simple [mlr3misc::Dictionary] storing objects of class [Optimizer]. Each
-#' optimizer has an associated help page, see `mlr_optimizer_[id]`.
+#' A simple [mlr3misc::Dictionary] storing objects of class [Optimizer].
+#' Each optimizer has an associated help page, see `mlr_optimizer_[id]`.
 #'
-#' This dictionary can get populated with additional optimizer by add-on
-#' packages.
+#' This dictionary can get populated with additional optimizer by add-on packages.
 #'
-#' For a more convenient way to retrieve and construct optimizer, see
-#' [opt()]/[opts()].
+#' For a more convenient way to retrieve and construct optimizer, see [opt()]/[opts()].
 #'
 #' @section Methods:
 #' See [mlr3misc::Dictionary].
 #'
+#' @section S3 methods:
+#' * `as.data.table(dict)`\cr
+#'   [mlr3misc::Dictionary] -> [data.table::data.table()]\cr
+#'   Returns a [data.table::data.table()] with fields "key", "param_classes", "properties" and "packages" as columns.
+#'
+#' @family Dictionary
 #' @family Optimizer
 #' @seealso
 #' Sugar functions: [opt()], [opts()]
+#'
 #' @export
 #' @examples
-#' opt("random_search", batch_size = 10)
+#' as.data.table(mlr_optimizers)
+#' mlr_optimizers$get("random_search")
+#' opt("random_search")
 mlr_optimizers = R6Class("DictionaryOptimizer", inherit = Dictionary, cloneable = FALSE)$new()
 
 #' @export
-as.data.table.DictionaryOptimizer = function(x, ...) {
+as.data.table.DictionaryOptimizer = function(x, ..., extra_cols = character()) {
+  assert_character(extra_cols, any.missing = FALSE)
+
   setkeyv(map_dtr(x$keys(), function(key) {
     t = tryCatch(x$get(key),
       missingDefaultError = function(e) NULL)
@@ -33,12 +42,9 @@ as.data.table.DictionaryOptimizer = function(x, ...) {
       return(list(key = key))
     }
 
-    list(
-      key = key,
-      param_classes = list(t$param_classes),
-      properties = list(t$properties),
-      packages = list(t$packages),
-      man = t$man
+    c(
+      list(key = key, param_classes = list(t$param_classes), properties = list(t$properties), packages = list(t$packages)),
+      mget(extra_cols, envir = t)
     )
   }, .fill = TRUE), "key")[]
 }