+++ tests/testthat/test_TerminatorStagnation.R
@@ -18,3 +18,8 @@ test_that("TerminatorStagnation works with empty archive", {
   archive = Archive$new(ps(x = p_dbl()), ps(y = p_dbl(tags = "minimize")))
   expect_false(terminator$is_terminated(archive))
 })
+
+test_that("man exists", {
+  terminator = trm("stagnation")
+  expect_man_exists(terminator$man)
+})